Refactoring: moved game-agnostic evaluator to the proper package.
