+++ src/main/java/com/leokom/games/chess/player/legal/brain/normalized/TwoPliesEvaluator.java
@@ -2,7 +2,7 @@ package com.leokom.games.chess.player.legal.brain.normalized;
 
 import com.leokom.games.commons.engine.GameState;
 import com.leokom.games.commons.engine.GameTransition;
-import com.leokom.games.chess.player.legal.brain.common.GenericEvaluator;
+import com.leokom.games.commons.brain.GenericEvaluator;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.ThreadContext;
