+++ src/test/java/com/leokom/games/chess/player/legal/brain/normalized/NormalizedBrainTest.java
@@ -2,7 +2,7 @@ package com.leokom.games.chess.player.legal.brain.normalized;
 
 import com.leokom.games.commons.engine.GameStateImpl;
 import com.leokom.games.commons.engine.GameTransitionImpl;
-import com.leokom.games.chess.player.legal.brain.common.GenericEvaluator;
+import com.leokom.games.commons.brain.GenericEvaluator;
 import org.hamcrest.CoreMatchers;
 import org.junit.Test;
